fixing a mismatch in TechTreeWnd between tech status assumptions and actuality, which sometimes led to crash, plus ensuring that TechList starts out populated (at expense of Reset() with every swap from layout to list)
Also minor tweaks in OptionsWnd and MapWnd to allow compilation on Linux, as previously mentioned in ObjectMap thread -- a couple spots casting double(), one spot casting std::string()

svn path=/trunk/FreeOrion/; revision=5887

+++ UI/MapWnd.cpp
@@ -3024,7 +3024,7 @@ void MapWnd::DoFleetButtonsLayout() {
 
 std::pair<double, double> MapWnd::MovingFleetMapPositionOnLane(const Fleet* fleet) const {
     if (!fleet) {
-        return std::make_pair<double, double>(UniverseObject::INVALID_POSITION, UniverseObject::INVALID_POSITION);
+        return std::make_pair<double, double>(double(UniverseObject::INVALID_POSITION), double(UniverseObject::INVALID_POSITION));
     }
 
     // get endpoints of lane on screen, store in UnorderedIntPair which can be looked up in MapWnd's map of starlane endpoints
